fix: wrap two unguarded np.clip() scalars in float() to stop float64 promotion (#5106)
np.clip(python_float, lo, hi) returns a numpy.float64 scalar, not a Python
float. Under NEP 50 that's a strong dtype, so multiplying a float32 audio
array by it silently promotes the whole result to float64 -- doubling
memory/CPU for every downstream array with no audible effect (the
project's dtype-in-[float32, float64] invariant still holds either way).

apply_reference_matching()'s gain_factor and _apply_simple_eq_fallback()'s
bass/treble gain_linear were the sole unguarded scalar-clip-into-audio
sites in auralis/core/ and auralis/dsp/ (re-verified via grep -- every
other np.clip() call either clips an array directly, which preserves
dtype by construction, or produces a control/coordinate value never
multiplied into an audio buffer). Wrapped both at the point of
computation, matching the existing float()-wrapped siblings in
harmonic_exciter.py, bass_enhancement.py, and dsp/utils/adaptive.py.
9th/10th instance of this pattern (#2158, #2450, #3468, #3659, #3687,
#4105, #4107, #4934/#4225).

Added a regression test verifying both functions preserve float32 through
their previously-promoting branches, plus a unit test pinning the fix at
its source (the clipped gain itself must be a plain float). Confirmed via
a clean-baseline worktree that all 4 dtype assertions genuinely fail
pre-fix (float64 instead of float32).

Tested: tests/auralis/core/test_clip_scalar_dtype_5106.py +
tests/auralis/dsp (206 passed); mypy clean on both changed files.

+++ auralis/core/processing/eq_processor.py
@@ -297,13 +297,15 @@ class EQProcessor:
         bass_gain_db = targets.get("bass_boost_db", 0.0)
         treble_gain_db = targets.get("treble_enhancement_db", 0.0)
 
-        # Apply very gentle adjustments (limited for safety)
+        # Apply very gentle adjustments (limited for safety). float()-wrapped:
+        # np.clip() returns a numpy.float64 scalar, which under NEP 50 would
+        # promote the float32 `processed` array it's multiplied into (#5106).
         if abs(bass_gain_db) > 0.5:
-            bass_gain_linear = 10 ** (np.clip(bass_gain_db, -6, 6) / 20)
+            bass_gain_linear = float(10 ** (np.clip(bass_gain_db, -6, 6) / 20))
             processed = processed * (1.0 + (bass_gain_linear - 1.0) * 0.1)
 
         if abs(treble_gain_db) > 0.5:
-            treble_gain_linear = 10 ** (np.clip(treble_gain_db, -6, 6) / 20)
+            treble_gain_linear = float(10 ** (np.clip(treble_gain_db, -6, 6) / 20))
             processed = processed * (1.0 + (treble_gain_linear - 1.0) * 0.1)
 
         return processed
